mbuto: support building a minimal image for kernel selftests
Add suboption -C for the case of -p kselftests for choosing a test
collection to run. For example, specifying "-p kselftests -C net"
creates an image that contains only the tests from the net collection
and only the kernel modules needed to run those tests.

If there are missing modules for running the tests, a warning will be
emitted showing missing modules names. The -C option is similar to
the -c option of run_kselftests.sh, but unlike run_kselftest.sh's
option it currently doesn't support specifying multiple collections.
This will be fixed later.

Also add suboption -T for the case of -p kselftests for choosing a
specific test. For example, specifying:

  -p kselftests -T net:ip_defrag.sh"

create an image that runs only the single test when booted. This
option is also similar to run_kselftest.sh's -t option, but it
lacks some functionality, which will be added later.

Signed-off-by: Sevinj Aghayeva <sevinj.aghayeva@gmail.com>

v1 -> v2: Update TOOLS, Use variable names for commands, and use capitalized
versions of run_kselftests.sh options (-C and -T).