mbuto: enable passing multiple collections and tests to suboptions
The run_kselftest.sh script, which is run upon booting the image created by
mbuto, allows specifying multiple collections and specific tests, as in:

    run_kselftest.sh -c "net proc" -t "x86:iopl_32 timers:rtcpie"

The above command runs (1) all the tests in the net collection, (2) all the
tests in the proc collection, (3) the iopl_32 test from the x86 collection,
and (4) the rtcpie test from the timers collection. This patch enables the same
functionality in mbuto. Now users can run mbuto as follows:

    mbuto -p kselftests -C "net proc" -T "x86:iopl_32 timers:rtcpie"

and mbuto generates an image that runs the same set of tests. In addition, mbuto
installs only the tests in the specified collections in the image. For example,
for the above command line, mbuto will install all the tests in the net, proc,
x86, and timers collections (but it will only run iopl_32 and rtcpie test from
the last two). Mbuto also installs only the modules needed for running these
tests and warns the user if there are missing modules.

Mbuto also warns the user if an invalid collection name is passed to the -C
suboption. Due to irregular structure of Makefiles in selftest directories, it
is hard to warn the user if an invalid test name is passed to the -T option. In
this case the user will find out about the mistake after the kernel boots and
run_kselftest.sh fails to run the test.

Signed-off-by: Sevinj Aghayeva <sevinj.aghayeva@gmail.com>